+++ skel/test/valeev.hpp
@@ -9,13 +9,13 @@ void Valeev_Init(void);
 void Valeev_Finalize(void);
 
 // As original as can be
-void Valeev_F(double *F, int n, double x);
+void Valeev_F(long double *F, int n, long double x);
 
-double Valeev_eri(int l1, int m1, int n1, double alpha1,
-                  const double* A, int l2, int m2, int n2,
-                  double alpha2, const double* B, int l3, int m3,
-                  int n3, double alpha3, const double* C, int l4,
-                  int m4, int n4, double alpha4, const double* D,
+long double Valeev_eri(int l1, int m1, int n1, long double alpha1,
+                  const long double* A, int l2, int m2, int n2,
+                  long double alpha2, const long double* B, int l3, int m3,
+                  int n3, long double alpha3, const long double* C, int l4,
+                  int m4, int n4, long double alpha4, const long double* D,
                   int norm_flag);
 
 #endif
